feat: adiciona trait modelo de oportunidade e separa do controller
