+++ src/core/Controllers/Opportunity.php
@@ -33,6 +33,7 @@ class Opportunity extends EntityController {
         Traits\ControllerArchive,
         Traits\ControllerAPI,
         Traits\ControllerAPINested,
+        Traits\EntityOpportunityDuplicator,
         Traits\ControllerEntityActions {
             Traits\ControllerEntityActions::PATCH_single as _PATCH_single;
         }
@@ -1292,113 +1293,5 @@ class Opportunity extends EntityController {
             }
         }
         $this->json($opportunity);
-    }
-
-    function ALL_duplicate(){
-        $app = App::i();
-
-        $this->requireAuthentication();
-        $opportunity = $this->requestedEntity;
-        $newOpportunity = clone $opportunity;
-
-        // duplica a oportunidade
-        $dateTime = new \DateTime();
-        $now = $dateTime->format('d-m-Y H:i:s');
-        $newOpportunity->setName("$opportunity->name  - [Cópia][$now]");
-        $newOpportunity->setStatus(Entity::STATUS_DRAFT);
-        $app->em->persist($newOpportunity);
-        $app->em->flush();
-
-
-        // duplica o método de avaliação para a oportunidade primária
-        $evaluationMethodConfigurations = $app->repo('EvaluationMethodConfiguration')->findBy([
-            'opportunity' => $opportunity
-        ]);
-        foreach ($evaluationMethodConfigurations as $evaluationMethodConfiguration) {
-            $newMethodConfiguration = clone $evaluationMethodConfiguration;
-            $newMethodConfiguration->setOpportunity($newOpportunity);
-            $newMethodConfiguration->save(true);
-
-            // duplica os metadados das configurações do modelo de avaliação
-            foreach ($evaluationMethodConfiguration->getMetadata() as $metadataKey => $metadataValue) {
-                $newMethodConfiguration->setMetadata($metadataKey, $metadataValue);
-                $newMethodConfiguration->save(true);
-            }
-        }
-
-        // Duplica as fases
-        $phases = $app->repo('Opportunity')->findBy([
-            'parent' => $opportunity
-        ]);
-        foreach ($phases as $phase) {
-            if (!$phase->getMetadata('isLastPhase')) {
-                $newPhase = clone $phase;
-                $newPhase->setParent($newOpportunity);
-
-                // duplica os metadados das fases
-                foreach ($phase->getMetadata() as $metadataKey => $metadataValue) {
-                    if (!is_null($metadataValue) && $metadataValue != '') {
-                        $newPhase->setMetadata($metadataKey, $metadataValue);
-                        $newPhase->save(true);
-                    }
-                }
-
-                $newPhase->save(true);
-
-                // duplica os modelos de avaliações das fases
-                $evaluationMethodConfigurations = $app->repo('EvaluationMethodConfiguration')->findBy([
-                    'opportunity' => $phase
-                ]);
-                foreach ($evaluationMethodConfigurations as $evaluationMethodConfiguration) {
-                    $newMethodConfiguration = clone $evaluationMethodConfiguration;
-                    $newMethodConfiguration->setOpportunity($newPhase);
-                    $newMethodConfiguration->save(true);
-
-                    // duplica os metadados das configurações do modelo de avaliação para a fase
-                    foreach ($evaluationMethodConfiguration->getMetadata() as $metadataKey => $metadataValue) {
-                        $newMethodConfiguration->setMetadata($metadataKey, $metadataValue);
-                        $newMethodConfiguration->save(true);
-                    }
-                }
-            }
-        }
-
-        foreach ($opportunity->getMetadata() as $metadataKey => $metadataValue) {
-            if (!is_null($metadataValue) && $metadataValue != '') {
-                $newOpportunity->setMetadata($metadataKey, $metadataValue);
-            }
-        }
-        
-        $newOpportunity->setTerms(['area' => $opportunity->terms['area']]);
-        $newOpportunity->saveTerms();
-
-        foreach ($opportunity->getRegistrationFieldConfigurations() as $registrationFieldConfiguration) {
-            $fieldConfiguration = clone $registrationFieldConfiguration;
-            $fieldConfiguration->setOwnerId($newOpportunity->getId());
-            $fieldConfiguration->save(true);
-        }
-
-        foreach ($opportunity->getRegistrationFileConfigurations() as $registrationFileConfiguration) {
-            $fileConfiguration = clone $registrationFileConfiguration;
-            $fileConfiguration->setOwnerId($newOpportunity->getId());
-            $fileConfiguration->save(true);
-        }
-
-        foreach ($opportunity->getMetaLists() as $metaList_) {
-            foreach ($metaList_ as $metaList__) {
-                $metalist = $metaList__;
-                $metalist->setOwner($newOpportunity);
-            
-                $metalist->save(true);
-            }
-        }
-
-        $newOpportunity->save();
-
-        if($this->isAjax()){
-            $this->json($opportunity);
-        }else{
-            $app->redirect($app->request->getReferer());
-        }
     }
 }
